multimodal-rag/requirements.txt
google-genai>=1.0.0
chromadb>=0.6.0
streamlit>=1.40.0
python-dotenv>=1.0.0
